+++ ucrypt_password.h
@@ -5,6 +5,6 @@
 #ifndef __UCRYPT_PASSWORD_H__
 #define __UCRYPT_PASSWORD_H__
 
-unsigned short int read_password(char* buffer,int *pass_len);
+unsigned short int read_password(char* buffer,unsigned short int *pass_len);
 
 #endif
